% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{bigger_than_10}
\alias{bigger_than_10}
\title{Bigger than 10
Tests if a number is bigger than 10.}
\usage{
bigger_than_10(n)
}
\arguments{
\item{n}{}
}
\value{
logical
}
\description{
Bigger than 10
Tests if a number is bigger than 10.
}
