% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{dependent_function}
\alias{dependent_function}
\title{Shows a dependency and pipe!}
\usage{
dependent_function(n_cyl)
}
\arguments{
\item{n_cyl}{number of cylinders}
}
\value{
filtered mtcars
}
\description{
Shows a dependency and pipe!
}
\examples{
dependent_function(x = mtcars, n_cyl = 4)
}
