% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{example_function}
\alias{example_function}
\title{My awesome first function}
\usage{
example_function()
}
\value{
always gives back 3
}
\description{
My awesome first function
}
\examples{
example_function()
}
