% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{give_me_y}
\alias{give_me_y}
\title{Give me the value of Y}
\usage{
give_me_y(x, n)
}
\arguments{
\item{x}{is a number}

\item{n}{the number to multiply by}
}
\value{
x * n
}
\description{
Give me the value of Y
}
\examples{
give_me_y(x = 5)
}
